hooked up the modules

+++ osgi-presentation/order/src/order/OrderService.java
@@ -1,6 +1,6 @@
 package order;
 
 public interface OrderService {
-	Receipt placeOrder(Long itemId,Long qty);
+	Receipt placeOrder(Long itemId,int qty);
 	void cancelOrder(Long orderId);
 }
